fix(comment diff): Show correct diff position
The existing code was using "StartRange", but the diff display algorithm
appears to do a better job if passed the "EndRange". Beyond that, we try to
cope with the API giving us different line positions if/when the position of
the note has shifted w/i the diff since it was first created.